+++ test/group.rs
@@ -1,4 +1,4 @@
-fn main() -> i64 {
+fn main() -> i32 {
     println(2 * (3 + 3));
     println(2 * 3 + 3);
     return 0;
